job-aggregator/requirements.txt
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
feedparser==6.0.11
playwright==1.44.0
anthropic==0.27.0
PyYAML==6.0.1
python-dotenv==1.0.1
pytest==8.2.0
pytest-mock==3.14.0
